% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readWrite_multiThreaded.R
\name{mcloadEnvironment}
\alias{mcloadEnvironment}
\title{mcloadEnvironment}
\usage{
mcloadEnvironment(loadpath = '/path/to/mcsaveEnvironment/directory', cores = 10)
}
\arguments{
\item{loadpath}{String. Path to a folder where the output of \code{mvsaveEnvironment()} is stored.}

\item{cores}{Integer. Default: 10. How many cores should be used for saving. Should be a value between 1 and 10.}
}
\description{
Load an R environment that was saved using \code{mcsaveEnvironment()}.
}
