% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readWrite_multiThreaded.R
\name{mcreadRDS}
\alias{mcreadRDS}
\title{mcreadRDS}
\usage{
mcreadRDS(file, cores = 4)
}
\arguments{
\item{file}{String. Path where the RDS file needs to be stored.}

\item{cores}{Integer. Default = 4. How many cores to use for saving. Typically setting this higher than 4 does not yield better performance.}
}
\value{
R-object
}
\description{
This is a multicore version of the default \code{readRDS()} function.
It requires a significantly shorter time to read RDS files.
It uses rapidgzip to decompress RDS files and load them into your environment.
Typically the performance does not increase anymore when using more than 4 cores.
When cores = 1, this functions uses the default \code{readRDS()} function.
}
\examples{
\dontrun{
myObj <- mcreadRDS('myObj.RDS')
}

}
