% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readWrite_multiThreaded.R
\name{mcsave.image}
\alias{mcsave.image}
\title{mcsave.image}
\usage{
mcsave.image(file = 'myEnvironment.RData')
}
\arguments{
\item{file}{String. Path to Rdata file where the environment needs to be saved.}

\item{compression_level}{Integer. Default = 3. Should be a value between 0 and 3.}

\item{cores}{Integer. Default = 4. How many cores to be used. Usually more than 4 cores does not yield much performance gains.}
}
\description{
This is a multicore version of save.image().
It uses igzip that automatically compresses the RData file.
It is not faster perse than \code{save.image()} without compression, but this function automatically compresses the RData file in the about same time as \code{save.image()} without compression.
}
