% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readWrite_multiThreaded.R
\name{mcsaveEnvironment}
\alias{mcsaveEnvironment}
\title{mcsaveEnvironment}
\usage{
mcsaveEnvironment(savepath = '/path/to/empty/directory', cores = 10)
}
\arguments{
\item{savepath}{String. Path to where to save the RData files. If it does not exists, it will be created. Best to use an empty or non-existing path.}

\item{save_functions}{Default: TRUE. TRUE/FALSE. Whether to store functions.}

\item{cores}{Integer. Default: 10. How many cores should be used for saving. Should be a value between 1 and 10.}

\item{compress}{TRUE/FALSE. Default:FALSE. Whether to use compression. This can make the files about half the size, but requires more time.}

\item{compression_level}{Integer. Default: 3. Should be a value between 0 and 3. Higher compression means smaller files, but also requires more time. Only relevant when \code{compress = TRUE}.}
}
\description{
Save an R environment using multiple compute cores.
This function is faster than the default \code{save.image()} and \code{mcsave.image()}.
Instead of saving the environment in a single .RData file (e.g. as is done with the default function \code{save.image()}), this creates a folder where each object, variable and function is stored as a separate file.
Each of these RData files can loaded in individually using \code{load()}, or the entire environment can be loaded from this folder using \code{mcloadEnvironment()}.
This functions automatically loads the parallel package.
}
