% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readWrite_multiThreaded.R
\name{mcsaveRDS}
\alias{mcsaveRDS}
\title{mcsaveRDS}
\usage{
mcsaveRDS(object, file, compression_level = 3, cores = 4)
}
\arguments{
\item{object}{Variable, object or function to be saved.}

\item{file}{String. Path where the RDS file needs to be stored.}

\item{compression_level}{Integer. Default = 3. Value between 1 and 3. Higher values means better compression, but also requires more time.}

\item{cores}{Integer. Default = 4. How many cores to use for saving. Typically setting this higher than 4 does not yield better performance.}
}
\description{
This is a multicore version of the default \code{saveRDS()} function.
It requires a significantly shorter time to save RDS files.
It uses igzip which automatically compresses the RDS files.
Typically the performance does not increase anymore when using more than 4 cores.
When cores = 1, this functions uses the default \code{saveRDS()} function.
}
\examples{
\dontrun{
mcsaveRDS(myObj, file = 'myObj.RDS')
}

}
